+++ Zora.Statistics/Models/Statistics/StatisticsViewModel.cs
@@ -4,8 +4,8 @@
     {
         public int TotalStudents { get; set; }
 
-        public double TotalPaidAmount { get; set; }
+        public decimal TotalPaidAmount { get; set; }
 
-        public double TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
     }
 }
